Moved Project over to Angular. READ DESCRIPTION!!!!
Don't Worry Angular isn't hard you won't have to learn anything...

To use type in CMD:

npm install -g @angular/cli

After that open the project and use the terminal to open the webpage:

ng serve --o

That should open the website for you. DONE

The PHP files are in: \src\assets\PHP

The HTML and SCSS Files are in Components (Seperated in different pieces)
locate them in: \src\app > Footer or the other ones. You can edit the files there if you need to.